data-fetcher-function/requirements.txt
google-cloud-firestore
googlemaps
requests